stream-processor/pyproject.toml
[project]
name = "stream-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "delta-spark>=4.0.0",
    "minio>=7.2.16",
    "pyspark>=4.0.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.12",
]
